[#182] refactor : Organize files under utils folder
	src/utils 폴더 아래 파일들을 각각의 카테고리에 맞게 정리했습니다.
