+++ src/services/internal/party/http.ts
@@ -4,7 +4,7 @@ import { JJAN_URL } from "../domain";
 import { Response } from "../types";
 
 import { httpService } from "@/module/http";
-import { pathToUrl } from "@/utils/pathToURL";
+import { pathToUrl } from "@/utils/route";
 
 export const joinParty = (partyId: string | undefined) =>
   httpService.post<Response<null>>(
