+++ src/utils/route/index.ts
@@ -1,0 +1,1 @@
+export { pathToUrl } from "./pathToURL";
